pyproject.toml
[tool.poetry]
name = "tmwbot"
version = "0.1.0"
description = ""
authors = ["verhovv <vladimirverhov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "3.0.0b7"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
pip = "^23.1.2"
python-dotenv = "^1.0.0"
fake-headers = "^1.0.2"
pandas = "^2.0.3"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
